portmaster_api_http::wire::factory::renderable

Trait Renderable

pub(crate) trait Renderable: Send {
    // Required methods
    fn write_flatbuffer(
        &self,
        builder: &mut Builder,
    ) -> Result<Offset<()>, ApiError>;
    fn write_json(&self, out: &mut Vec<u8>) -> Result<(), ApiError>;
}

Uma resposta que sabe se escrever, sem dizer de que tabela veio.

Por que existem duas traits de resposta

Offset<T> não faz upcast: o planus oferece downcast() para Offset<()> e nada no sentido contrário. Uma strategy que segurasse &dyn ResponseFactory não teria como chamar table(), porque o tipo associado não sobrevive ao apagamento.

Daí o par: ResponseFactory é tipada, para o pai aninhar o filho; esta é apagada, para a strategy segurar num Box. O blanket impl abaixo liga as duas, e nenhuma factory precisa saber que esta trait existe.

Os bytes são idênticos aos que sairiam pelo caminho tipado: Offset<T>::ALIGNMENT é 4 para todo T, e o file_identifier é None nos dois casos.

Required Methods

fn write_flatbuffer( &self, builder: &mut Builder, ) -> Result<Offset<()>, ApiError>

Escreve a tabela no builder e devolve o offset já sem tipo.

fn write_json(&self, out: &mut Vec<u8>) -> Result<(), ApiError>

Serializa a tabela como JSON no buffer dado.